Add files via upload
A finite element analysis of the robot structure was performed concerning the value of the theoretical polishing force. Figures respectively show the total deformation of the robot (the maximum value is 0.0403 mm) and the equivalent elastic strain ( The maximum value is 1.0117e-5 mm/mm), which proves that the robot is safe and stable in this working state. （The Ansys file is too large, exceeding 50G, as shown in the screenshot file data.）